.get & query First Introduction
11.1.5
The start of using .send to test if things work and then .get to actually grab the objects array from the animals.json.  Then we set up filterByQuery function and called it in the .get so when we use the browser looks like this --http://localhost:3001/api/animals?name=Erica-- we can get Erica's object by her name.  Very dense.  May need to read agian.