fs and path to add json to animals.json
11.2.6
Created a function to handle adding a new animalsArray. Then we required fs and path to make it so we can write the files and convert them to json to add to the animals.json file. Finally we added some validation to make sure everything in the object appears before we add it to the page. If just one item is missing we spit out an error and do not add to the page.