sendFile and express.static
11.3.4
Used app.get with sendFile(path.join) to go get the index.html in the public folder and display it to the browser.  Then we used an express function --app.use(express.static("public"));-- to make sure all files in the "public" folder are referenced each time the page is deployed.