New Fetch to display animal boxes on the page.
11.3.7
Added a new fetch that went to our server.js which then went into animals.json and got the info we needed to diplay to the page. Then it used the printResults() already built for us to actually create the boxes to display it to the page. I may have to reread this one so I can understand it better.