ROUTER & File separation
11.4.4
Introduced to ROUTER and separated a ton of things out of the server.js.  Really dense stuff that will have to referenced to understand.  But essentially ROUTER lets us use express() everywhere without constantly creating NEW express calls.  It can work with the existing call and send it through several files.  The require looks like this --const router = require('express').Router();-- we change all the app.get to router.get etc.. and we export it like this --module.exports  = router;--.